moved musical notes to module
The code that generates the creates the musical frequencies & maps the
keyboard keys to specific frequencies has been moved to a separate
module in order to de-clutter the scripts.js file.
